Add files via upload
This is a short program to determine the results of an election written in C. I was tasked with writing the "vote" and "print_winner" functions. For the "print_winner" function in particular, I wrote a short custom algorithm that looped the list 1 time to gather the highest vote count, and then looped the list 1 more time to add the candidates who possessed the highest vote count to an array. This is an O(n) time complexity algorithm. I then outputted the names of the highest voted for candidates.